Make loading indicator show based on stack
This means that we keep track of how many things want to show a loading
indicator, and only hide it when all of those have reported back that we
can hide it. This makes for a nicer experience, because the loading
indicator doesn't disappear until _everything_ has loaded.

Also fixes a bug where the loading indicator would stay on indefinitely
after switching to a story one has seen before. Closes #267