Remove all recompose/pure usage
We have issues in the app where the favicon won't update, even if new
notifications come in. I have a hunch that this is because we're using
`recompose/pure` a lot throughout the app.

Not only that, but `recompose/pure` does a shallow comparison between
the current props and the next props, which can lead to subtle bugs. It
also often is more expensive to check the difference in props than to just
render again. (especially true for leaf components)

This commit goes ahead and removes all instances of `recompose/pure` usage.
Now we can introduce it again in places where we notice performance
problems, but only in those places and nowhere else. We need to use it like
surgeons but currently we're like a gatling gun throwing it anything
that moves.