WIP: Switch to offline-plugin for ServiceWorker caching
The `offline-plugin` has a lot more options in terms of how we want to
cache things so that we can get rid of those annoying white pages and
"SyntaxError: < is not a fucking fuck" errors.

Specifically, this let's us update the page when a user switches to a
separate page. This means they'll continue being on the old version, but the
ServiceWorker updates in the backgroud and downloads the new version.

Once that's done, a user clicks on a new page, but rather than
navigating them there client-side we do a full page reload so we render
with the new code.

That's based on this article:
https://zach.codes/handling-client-side-app-updates-with-service-workers/
which led me down this route.

This is a WIP, there's some stuff that we need to do before it's
ready to be shipped:

1. Implement cache pass-throughs for `/api`, `/auth` and maybe some
others
2. Figure out wether we should use a custom `Link` component or not
(i.e. test)
3. Make sure this works exactly as intedended.
4. Probably some other stuff, y'all know how it is

This also fixes the `webpackJsonp is not defined` error.