WIP: Track user lastSeen status on threads
This tracks a users lastSeen status on threads. I made the assumption
that when a user loads the messages of a thread is when they've read it.
I'm open to discussing this if you have better ideas, but I think it's
pretty valid to assume.

If you've read the discussion, is there anything left to do? I don't
think so.

There's still some stuff left to do:

- [ ] Update lastSeen when subscribed and new message comes in
- [ ] Use this on the frontend somehow (ideas @brianlovin @uberbryn?)