Don't retry queries for as long
Retrying queries for a long time is very annoying, as it leads to
showing the loading indicators the whole time. That makes for horrible
UX as the user doesn't even know they're disconnected.

Instead, we now only retry mutations for a long time (because we really
really want them to go through) and only retry queries for a couple
seconds at most.