Rework the asyncify utility
Based on [the code used in grapqhl-subscriptions to convert an
EventEmitter to an AsyncIterator](https://github.com/apollographql/graphql-subscriptions/blob/master/src/event-emitter-to-async-iterator.ts)
this patch reworks our `asyncify` util, which is used in all subscriptions.

Rather than having a queue of queues and a single database listener
(which obviously didn't work and led to memory leaks due to people
unsubscribing and the queue sticking around) we now listen to changes
_per user and subscriptions_. This allows us to open new listeners
whenever a user connects, and close the listener whenever a user
disconnects.

This should lead to much less memory problems, although it might lead to
database problems. (we need to test that in prod) If it does lead to
more frequent connection drops with RethinkDB I'll rework this again to
only use one database listener for all users.