Fix lastSeen injection bugs
By moving the lastSeen injection to the component, rather than doing in
in groupMessages, we can cache the lastSeen and threadLastActive values
so that the unseen robo doesn't jump around while looking at an active
thread.

This is a much better experience in general, and also happens to fix
thread with more than 50 messages in the inbox. Closes #2469