Fix ServiceWorker issue once and for all
This patch introduces new routing behavior for static JS files being
served from hyperion. No matter what the hash of the requested file is,
the request will be redirected to the proper JS. E.g. if you request
`/static/js/main.some-bullshit-thats-wrong.js` you will automatically be
redirected to `/static/js/main.asdf123.js`. (where `asdf123` is the
correct and up to date hash of the file on the disk)

This is necessary because of our client-side ServiceWorker caching. It
seems to happen fairly often that it requests a new file, but due to a
redeploy (and thusly the file hashes changing) the file isn't at that
place anymore-so hyperion tries to SSR that path and returns HTML.