Fix HMR error
We can't HMR src/index.js, but we can HRM src/routes.js. The issue is
that we don't want to HRM src/routes.js on the server, so we create
another file src/hot-routes.js which wraps the src/routes.js in the hot
reloading stuff which is only used on the client.