Fix perpetually being signed out AGAIN
The issue with being perpetually signed out turns out to be that the
Zeit CDN caches the homepage, and doesn't know whether users are logged
in or not, and thusly serves the logged-out homepage on refresh even
though one might be logged in! :scream:

This fixes it by setting a `__now_no_cache` cookie when signing in.